+++ src/axom/core/tests/core_memory_management.hpp
@@ -516,7 +516,7 @@ TEST(core_memory_management, interspace_reallocation)
   int* origOnHost = axom::allocate<int>(maxNK, axom::MALLOC_ALLOCATOR_ID);
   for(std::size_t i = 0; i < maxNK; ++i)
   {
-    origOnHost[i] = 100 + i;
+    origOnHost[i] = static_cast<int>(100 + i);
   }
   int* tempOnHost = axom::allocate<int>(maxNK, axom::MALLOC_ALLOCATOR_ID);
